northd: Support active-active lrps.
In case we find a LRP (or a LSP connected to a LRP) that has
options:active-active-lrp set we ignore it during normal processing in
join_logical_ports.

We add an additional section at the end where we then use these ports to
generate derived Port_Bindings for each LRP + LSP combination once for
each matching ovn-aa-port-mappings entry.

In the end this gives us the same result as if someone would have
precreated a LRP + LSP combination for each ovn-aa-port-mappings in the
northbound. However it allows our users to benefit from active-active
routing without their CMS needing to know about this feature (besides
the active-active-lrp setting).

Signed-off-by: Felix Huettner <felix.huettner@stackit.cloud>