WIP: Advertisement of NAT/LB addresses as routes.
This version attempts to split northd engine processing
of advertised routes. The main motivation is to avoid
logical flow recomputation when NAT/LB routes change.

+

WIP: Advertisement of NAT/LB pt.2.

This is a continuation of previous commit. It's separate
for ease of review. It will be squshed together for the
final version.

+

Review suggestions from dceara.

+

Not cleaning the dynamic_routes_data means that we just accumulate
entries over iterations, even though their references are no longer
valid.

Signed-off-by: Martin Kalcok <martin.kalcok@canonical.com>
Signed-off-by: Felix Huettner <felix.huettner@stackit.cloud>