northd: Check and populate NAT entry fields early in the I-P engine.
We can do this early, in the en_lr_nat node.  That allows future commits
to reuse the parsed information later in the I-P run without having to,
reprocess NB contents.

While at it clean up a bit the nat entry parsing and checking.  We were
re-parsing external_ips, for example, twice even though we already had
the parsed IPs available in the internal record.  Also the code now
tries to separate parsing and checking a bit as they were inter-twined
and very hard to understand.

Signed-off-by: Dumitru Ceara <dceara@redhat.com>