controller: Watch for (Linux) neighbor changes.
For each (Linux) interface that's interesting for the OVN control plane,
watch for neighbor (fdb/arp) changes.  This allows us to reconcile on
changed neighbor entries from outside routing agents (e.g., FRR).

This is the neighbor equivalent of the support added for route tables in
673d90f1173f ("controller: Watch for route changes.").

Acked-by: Xavier Simonart <xsimonar@redhat.com>
Signed-off-by: Dumitru Ceara <dceara@redhat.com>
---
V4:
- Addressed Ilya's comments:
  - sizeof, VLOG_DBG, typo

V3:
- renamed fdb_table_change() to neighbor_table_change()
- added tests

Changes in V2:
- fixed up log messages
- changed code to make OVN only manage neighbor entries with VLAN 0
- added advertise_neigh_find() helper