controller: Add host-if-monitor to track (Linux) interface indices.
When monitoring and updating (Linux) neighbor tables through Netlink,
the interfaces in question are identified by if-index.  It's more
CMS-friendly to allow OVN to be configured to track interfaces by name.

In order to achieve that, we introduce a new host-if-monitor module
which tracks the if-index <-> if-name mapping for relevant interfaces.

NOTE: OVS has the lib/if-notifier.[ch] module already.  However, that is
not really useful for the OVN use case because it doesn't really pass
the struct rtnetlink_change message to its registered callbacks.

Signed-off-by: Dumitru Ceara <dceara@redhat.com>