northd: Add an option to specify EVPN vni in logical switches.
Propagate the VNI option from LS to the datapath, this is the last
missing piece required to activate all the code around EVPN in
ovn-controller. The LS is considered to be part of EVPN when
the "dynamic-routing-vni" is set to valid integer.

Reported-at: https://issues.redhat.com/browse/FDP-1385
Acked-by: Xavier Simonart <xsimonar@redhat.com>
Signed-off-by: Ales Musil <amusil@redhat.com>
---
v4: Rebase on top of latest main.

v3: Rebase on top of latest main.
    Add Xavier's ack.

v2: Rebase on top of latest main.
    Add system test.
    Fix typo in documentation.