northd: Use complete matching for ipv6 LLA nexthops
The logical router port for outport of a logical router was calculated
based on the subnet matching of LRPs on the given LR. This doesn't work
in case IPv6 local-link addresses are used.

The patch introduces a generic function that calculates outport by
strict matching on the nexthop address. This function is used when
generating the logical flow from the logical router policy.

This problem becomes more complex when a router is connected to
complicated topology where the L2 domain is stretched over multiple
logical switches or even localnet ports. This complex solution is not
included in this patch and will be part of a followup patch.

Same problem seems to exist when finding the outport from the logical
router static route. The function responsible for finding the right LRP
is genereic enough so it can be re-used later for the static route too
in another followup patch.

Reported-at: https://issues.redhat.com/browse/FDP-1554
Reported-by: Jakub Libosvar <jlibosva@redhat.com>
Signed-off-by: Jakub Libosvar <jlibosva@redhat.com>