northd: datapath-sync: Add IP for LS and LR inputs.
Introduce Incremetal processing logic to datapath-sync node for
en_datapath_logical_switch and en_datapath_logical_router inputs.

When processing incrementally, the behavior of en-datapath-sync is
different than when run non-incrementally. When new northbound logical
datapath types are added, en-datapath-sync will not propagate the new
logical datapath downstream to other nodes. Instead, en-datapath-sync
waits for the southbound database to report the newly inserted
Datapath_Binding, and then en-datapath-sync reports the new datapath.
See the large comment in northd/datapath-sync.h for more details.

One test change that may seem unrelated is the propage sb_cfg test in
ovn-northd.at. That test checked for a specific sequence of events to
happen based on northbound database manipulation. The test had used the
insertion of a logical switch to test this. Now, inserting a logical
switch causes more than one engine run to occur because of the way
en-datapath-sync works. The easiest way to fix the test was to insert a
different type of row into the database, so this commit changes to
using a port group now.

Reported-at: https://issues.redhat.com/browse/FDP-1519
Signed-off-by: Lorenzo Bianconi <lorenzo.bianconi@redhat.com>