en-advertised-route-sync: Fix connected-as-host behavior for LLA-only ports.
Commit f2deb24c5c43 ("northd: Sync Advertised_Route to sb.") correctly
skipped over "link-local connected networks" not creating
Advertised_Route records for the link-local subnet.  However, when
implementing the connected-as-host semantics it shouldn't have skipped
over the (non-LLA) IPs configured on routers/switches connected to said
interfaces.

It introduced an inconsistency between the case (a) when a router port is
configured with an explicit IP (and network) and the case (b) when the
router port is only configured with MAC address (just a link-local IPv6
address).  In (a), if the router port is configured to advertise
"connected-as-host" ovn-northd creates Advertised_Route records for all
IPs known to be configured on the peer switch/router.  In (b), however
it does not.

This forces CMS to always configure an explicit IP on the router port
even though that's not necessary for any of the routing decisions.

Fix it by advertising the connected-as-host routes for all connected
routes generated for a router port.

NOTE: we don't try to verify that advertised host routes are actually
reachable (there's a route prefix on the router that matches on them).
This was already the ovn-northd behavior.

Reported-at: https://issues.redhat.com/browse/FDP-1563
Reported-by: Jakub Libosvar <jlibosva@redhat.com>
Fixes: f2deb24c5c43 ("northd: Sync Advertised_Route to sb.")
Signed-off-by: Dumitru Ceara <dceara@redhat.com>