northd: Handle SB records using Datapath_Bindings that were recreated on upgrade.
This is specifically needed for "dynamic" records created in the SB DB
by ovn-controller.  These are: MAC_Binding, IGMP_Group and
Learned_Route.

When Datapath_Bindings are recreated at upgrade (to move to the
new-style of synced NB->SB UUIDs) we need to ensure all "dynamic"
records are updated (so not removed) in order to avoid creating traffic
disruption until ovn-controller recreates them.

In order to achieve that we detect in the en-datapath-sync node when a
SB.Datapath_Binding will be deleted due to it being "old-style" and
store these into a map.  Once the "new-style" datapaths are created we
iterate through all types of records listed above and update any
references from "old" to "new" style datapath pointers.

While this might sound costly, in practice this will happen exactly
once for all datapaths at the same time, when the cluster is upgraded.

Fixes: 6919992d8781 ("Datapath_Binding: Separate type column and sync NB.UUID to SB.")
Reported-at: https://mail.openvswitch.org/pipermail/ovs-dev/2025-September/425918.html
Reported-by: Ilya Maximets <i.maximets@ovn.org>
Signed-off-by: Dumitru Ceara <dceara@redhat.com>