controller: Fix mapping of learned routes to LRP.
The commit in the fixes tag introduced options to map between the
system interface routes are learned on, and an LRP.

However, when an LR has LRPs with and without the option,
routes are attempted associated with LRPs without the option,
contrary to the documented behavior.

This was hidden in the original dynamic-routing - Gateway Router
test by populating all LRPs with the option, however this does
not represent a real world use case.

Ensure only LRPs with the option are considered when at least one
of LRPs in an LR has the option set.

Reported-at: https://launchpad.net/bugs/2123914
Fixes: d7d886eca553 ("controller: Support learning routes per iface.")
Signed-off-by: Frode Nordahl <fnordahl@ubuntu.com>
Signed-off-by: Dumitru Ceara <dceara@redhat.com>