Support explicitly specified output-port for router policies.
This closes the gap with the static routes support and allows users to
explicitly break the tie for cases that are ambiguous, e.g., when the
reroute next hop IP could be theoretically reachable on multiple
connected router ports.  An initial proposal tried to solve this problem
in a different way [0] but upon further discussions [1] we agreed to match
the support and static routes currently offer.

The NB router policy schema is extended to include an "output_port"
optional field which will be used by the CMS to explicitly determine
which router port to be used when the reroute policy is applied.  This
overrides any automatic port detection that happens for policies without
an "output_port" explicitly specified.

As mentioned in the documentation and in the TODO.rst file, the support
is limited to non-ECMP policies.  That's mainly because the current
router policy schema makes it extremely cumbersome to attach different
ports to different next hops that may be defined for the policy.  Unlike
NB static routes, the router policies store the ECMP nexthops as
comma-separated strings in the NB policy record.

[0] https://mail.openvswitch.org/pipermail/ovs-dev/2025-July/424894.html
[1] https://mail.openvswitch.org/pipermail/ovs-dev/2025-September/426231.html

Reported-at: https://issues.redhat.com/browse/FDP-1554
Reported-by: Jakub Libosvar <jlibosva at redhat.com>
Signed-off-by: Dumitru Ceara <dceara@redhat.com>