evpn-binding: Improve the evpn_datapath structure and optimize lookups.
Store the evpn_datapaths in I-P node data as it will be used by
upcoming patches, as input for new I-P nodes.  We also store a pointer
to 'struct local_datapath' objects instead of just the datapath key in
order to avoid looking up the datapath multiple time when reading the
(SB) EVPN datapath configuration is needed.  Instead of storing EVPN
datapaths as a vector, store them in a hashmap to avoid linear lookups.

Signed-off-by: Dumitru Ceara <dceara@redhat.com>