ci: ovn-kubernetes: Enable ovn-k UDN tests in our CI.
UDN stands for "User Defined Networks" and is a subset of the
ovn-kubernetes functionalities.  It's relevant for us to also
run some of the UDN tests because the internal UDN implementation
relies on some core OVN features that are not used by the other
ovn-kubernetes tests we run today.  For example, "conditional NAT"
and "transit router".

In order to do that we also need to:
- load the VRF kernel module in the ovn-kubernetes workflow
- add some missing env/workflow variables needed by the new E2E jobs
- bump E2E job timeouts to match what OVN-k uses

Reported-at: https://issues.redhat.com/browse/FDP-3032
Signed-off-by: Dumitru Ceara <dceara@redhat.com>