controller: Make sure we remove stale routes first.
In case we want to replace route with the same destination
but just different nexthop we need to remove the stale ones
first otherwise the first loop would return that the route
already exists and it would be created the second loop.
At the same time remove the unused err variable in
the re_nl_cleanup_routes function.

Signed-off-by: Ales Musil <amusil@redhat.com>