controller: Allow advertised routes to have next hop IP.
The advertised routes would always be specified as blackhole
without any nexthop. Allow the route to be configured with
specific nexthop even from different IP family to allow
routing of IPv4 over IPv6 nexthop.

At the same time also extend the netlink testing with
route-sync so we can test this out independently of
the rest of route advertisement.

Signed-off-by: Ales Musil <amusil@redhat.com>