controller: Make sure we don't do shallow copy of route_data.
The struct route_data includes an ovs_list, by copying it into vector
we might be copying soon to be invalid pointers, in other words
pointers to stack variables. This wasn't an issue until now because
we didn't reference anything within the list, however it has become
an issue with the nexthop reference added later on.

Fixes: 13d1a5cbae68 ("controller: Postpone the route deletion after the dump is done.")
Signed-off-by: Ales Musil <amusil@redhat.com>